[优化] comment接口优化
