+++ deploy/docker-compose/docker-compose.data.yaml
@@ -26,7 +26,7 @@ services:
       --collation-server=utf8mb4_general_ci #设置utf8字符集
 
   redis:
-    image: redis
+    image: redis:latest
     container_name: redis-server # 容器名
     restart: always
     ports:
@@ -37,7 +37,7 @@ services:
       - /home/docker/redis:/data
 
   rabbitmq:
-    image: rabbitmq
+    image: rabbitmq:management
     container_name: rabbitmq-server
     restart: always
     ports:
@@ -48,6 +48,6 @@ services:
       - RABBITMQ_DEFAULT_PASS=rabbitmq7914
     volumes:
       - /home/docker/rabbitmq:/var/lib/rabbitmq
-    command:
-      - rabbitmqctl status
-      - rabbitmq-plugins enable rabbitmq_management # 开启rabbitmq web管理界面
+#    command:
+#      - rabbitmqctl status
+#      - rabbitmq-plugins enable rabbitmq_management # 开启rabbitmq web管理界面
