+++ server/api/router/logic/register.rt.go
@@ -26,21 +26,6 @@ func (s *CategoryRouter) InitCategoryRouter(publicRouter *gin.RouterGroup, login
 	}
 }
 
-// 初始化 Comment 路由信息
-// publicRouter 公开路由，不登录就可以访问
-// loginRouter  登录路由，登录后才可以访问
-func (s *CommentRouter) InitCommentRouter(publicRouter *gin.RouterGroup, loginRouter *gin.RouterGroup) {
-	s.InitCommentBasicRouter(publicRouter, loginRouter)
-	var handler = s.svcCtx.CommentController
-
-	{
-		publicRouter.POST("comment/:id/like", handler.LikeComment)                // 点赞评论
-		publicRouter.POST("comment/:id/reply_list", handler.FindCommentReplyList) // 分页查询Comment列表
-		publicRouter.POST("comment/details_list", handler.FindCommentDetailsList) // 分页查询Comment列表
-		publicRouter.POST("comment/list/back", handler.FindCommentBackList)       // 分页查询Comment列表
-	}
-}
-
 // 初始化 Role 路由信息
 // publicRouter 公开路由，不登录就可以访问
 // loginRouter  登录路由，登录后才可以访问
