+++ server/api/router/logic/sp_website.rt.go
@@ -23,7 +23,7 @@ func (s *WebsiteRouter) InitWebsiteRouter(publicRouter *gin.RouterGroup, loginRo
 
 	var handler = s.svcCtx.WebsiteController
 	{
-		publicRouter.GET("", handler.GetBlogHomeInfo)                // 获取博客首页信息
+		publicRouter.GET("/", handler.GetBlogHomeInfo)               // 获取博客首页信息
 		publicRouter.GET("ws", handler.WebSocket)                    // websocket
 		publicRouter.POST("chat/records", handler.FindChatRecords)   // 查询前台聊天记录
 		publicRouter.GET("about/me", handler.GetAboutMe)             // 查询关于我
