+++ server/quickstart/quick_test.go
@@ -98,7 +98,6 @@ func TestPlate(t *testing.T) {
 	//gen.ApplyMetas(gen.GenerateMetasFromTable("api", "接口"))
 	//gen.ApplyMetas(gen.GenerateMetasFromTable("article", "文章"))
 	//gen.ApplyMetas(gen.GenerateMetasFromTable("category", "文章分类"))
-	//gen.ApplyMetas(gen.GenerateMetasFromTable("comment", "评论"))
 	//gen.ApplyMetas(gen.GenerateMetasFromTable("friend_link", "友链"))
 	//gen.ApplyMetas(gen.GenerateMetasFromTable("menu", "菜单"))
 	//gen.ApplyMetas(gen.GenerateMetasFromTable("operation_log", "操作记录"))
@@ -111,6 +110,9 @@ func TestPlate(t *testing.T) {
 	//gen.ApplyMetas(gen.GenerateMetasFromTable("talk", "说说"))
 	//gen.ApplyMetas(gen.GenerateMetasFromTable("chat_record", "聊天记录"))
 
+	// 不能覆盖的
+
+	//gen.ApplyMetas(gen.GenerateMetasFromTable("comment", "评论"))
 	//gen.ApplyMetas(gen.GenerateMetasFromTable("user_account", "用户账号信息"))
 	//gen.ApplyMetas(gen.GenerateMetasFromTable("user_information", "用户信息"))
 	//gen.ApplyMetas(gen.GenerateMetasFromTable("user_login_history", "用户登录历史"))
